changed: re-implemented the way `progmem` works.
This commit drastically changes how progmem is implemented, mainly
introducing type-checking so that only value types can be tagged with
the enum macro-pragma.

- The `ProgramMemory` type, which is the underlying type of any
variable tagged with `{.progmem.}` is not exported anymore, acting as
an opaque type.

- Renaming various `importc` stuff to pm_* for code readability.

- Solved a "bug" where progmem dereferences where creating temporary
elements instead of performing the copy directly onto the destination.
Now `[]` and `readFromAddress` are implemented as templates enforcing
non initialization of the result.

- `progmem` is able to typecheck the statements tagged with it: using
it with any non-value type (or even with managed value types like
`string` or `seq`) will give you a compilation error. This includes a
and much better object/array solver when translating nim -> progmem c.